Major overhaul to RAID system. Should not function correctly.

+++ mods/scourgecraft/tileentity/TileEntityGoldStorage.java
@@ -49,11 +49,11 @@ public class TileEntityGoldStorage extends TileEntityScourgeResource
 		return 0;
 	}
 	
-	public int storeGold(int amount)
+	public double storeGold(double amount)
 	{
 		if (amount + getGold() >= getMaxStorage())
 		{
-			int amountLeft = (int)getMaxStorage() - getGold();
+			double amountLeft = getMaxStorage() - getGold();
 			amountLeft = amount - amountLeft;
 			return amountLeft;
 		}
